Rename selection helpers
