+++ app/assets/javascripts/components/annotations/selectionHelpers.ts
